Add procedure to migrate the database, but we need to test it.
In addition, this commint introduce a important TODO that is from the db scheduler, we need to mantains the state of the UI in the setting UI. This means that we need to check what is the type of scheduler.

Maybe we need to introduce some propriety, to flag a manual propriety. Let see.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>